Clients public IPs are served to DNS
Upon signing up, the server and client
figure out which IP's the server can connect to,
and as a result, which IP's are assumed to be
publicly accessible.

These IPs are stored, and returned as the result
of DNS requests, rather than the signalling channel
IP, which was previously returned, but should not
be used for anything but signalling.

This squashed commit contains a whole range of
commits from both Haris and Sebastian.