Simple tactic for direct connections
This tactic doesn't work very well at the moment.
It still highlights how tactcs can be written,
and how the system currently operates.

It has each of the nodes get a list of the ips it sees itself having.
Then each of the nodes try to send datagrams to the other nodes ips.
Each node then returns the IPs it received datagrams from...

Needs more testing, and the results returned by the tactic are
currently only printed to the stdout, rather than added to the nodes
themselves.