Stores and returns public ip-addresses
Upon receiving a hello message from a client,
the server sends an RPC queryring the client for its
ip-addresses. The server then tries to send a UDP
packet to each of the addresses, and collectively with
the client discovers which addresses are publicly
available and which ones are not.
The ones that are publicly available are then stored in
the connections table, and returned alongside other
tunnels as the answer to DNS requests.