Reworked RPC module
The RPC module did not comply with JSON-RPC, and had
grown somewhat out of control.
In this commit is it trimmed down to support JSON-RPC
as specified on http://json-rpc.org/wiki/specification.
The parser is very strict and unforgiving. If messages
do not conform with the standard, or if there are
ambiguities, then they will get rejected.

Tactic specific RPCs are now also encoded as standard
JSON-RPCs, and the OpenVPN connection module and the
OpenVPN Manager have been updated accordingly.

This commit also adds functionality that allows
the signpost client setup script autodetect
the server settings, so a user does not have to
manually type them.