Fixed notification parsing bug
Notifications were parsed as if the ID field
should not exist, while the JSON-RPC specifies
them as being required to be Null. The creation
of the Notification messages was correct, but not
the parsing.

The notification also tried to parse for "args"
rather than "params".
Sending and receiving requests, responses and notifications
now seem to work.