Make direct connection tactic work
The direct connection tactic now works on
my local box. The nodes that are trying to
connect send UDP messages to one another, and
notice which IPs they can receive data on.

This data is then stored in the connections
module along with the status of the lookup.

When trying to connect to nodes, the engine
now takes into account which tactics have
already been run, in order to avoid rerunning
them repeatedly.

The dev box transfer now does a full transfer when it is
started, in case the state has changed on the dev machine
when the program was not running